+++ crates/common/src/tree/proofs.rs
@@ -1,0 +1,136 @@
+use anyhow::{Context, Result};
+use bincode;
+use jmt::{
+    proof::{SparseMerkleProof, UpdateMerkleProof},
+    KeyHash, RootHash,
+};
+use serde::{Deserialize, Serialize};
+use std::convert::Into;
+
+use crate::{
+    digest::Digest,
+    hashchain::{Hashchain, HashchainEntry},
+    hasher::Hasher,
+};
+
+#[derive(Serialize, Deserialize)]
+/// Represents a contiguous stream of [`Proof`]s leading from [`Batch::prev_root`] to [`Batch::new_root`].
+/// Used as the input to the circuit.
+pub struct Batch {
+    pub prev_root: Digest,
+    pub new_root: Digest,
+
+    pub proofs: Vec<Proof>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum Proof {
+    Update(Box<UpdateProof>),
+    Insert(Box<InsertProof>),
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+/// Represents an insertion proof for a newly created hashchain.
+/// Currently, this proof is generated by
+/// [`crate::operation::Operation::CreateAccount`] and
+/// [`crate::operation::Operation::RegisterService`] operations.
+// TODO(CRITICAL): Verify service's challenge input via a merkle proof of
+// the service, and then signature verification with the contained VK.
+pub struct InsertProof {
+    /// Proof that the key does not already exist in the tree (i.e. it's not overwriting an existing key)
+    pub non_membership_proof: NonMembershipProof,
+
+    /// Post-insertion root hash of the tree
+    pub new_root: Digest,
+    /// Proof that the new hashchain is correctly inserted into the tree
+    pub membership_proof: SparseMerkleProof<Hasher>,
+
+    /// The new hashchain entry that was inserted.
+    pub new_entry: HashchainEntry,
+}
+
+impl InsertProof {
+    /// The method called in circuit to verify the state transition to the new root.
+    pub fn verify(&self) -> Result<()> {
+        self.non_membership_proof.verify().context("Invalid NonMembershipProof")?;
+
+        let hashchain = Hashchain::from_entry(self.new_entry.clone())?;
+        let serialized_hashchain = bincode::serialize(&hashchain)?;
+
+        self.membership_proof.clone().verify_existence(
+            self.new_root.into(),
+            self.non_membership_proof.key,
+            serialized_hashchain,
+        )?;
+
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+/// Represents an update proof for an existing [`Hashchain`], updating it with a new [`HashchainEntry`].
+pub struct UpdateProof {
+    pub old_root: RootHash,
+    pub new_root: RootHash,
+
+    pub key: KeyHash,
+    pub old_hashchain: Hashchain,
+    pub new_entry: HashchainEntry,
+
+    /// Inclusion proof of [`UpdateProof::old_hashchain`]
+    pub inclusion_proof: SparseMerkleProof<Hasher>,
+    /// Update proof for [`UpdateProof::key`] to be updated with [`UpdateProof::new_entry`]
+    pub update_proof: UpdateMerkleProof<Hasher>,
+}
+
+impl UpdateProof {
+    /// The method called in circuit to verify the state transition to the new root.
+    pub fn verify(&self) -> Result<()> {
+        // Verify existence of old value.
+        // Otherwise, any arbitrary hashchain could be set as old_hashchain.
+        let old_serialized_hashchain = bincode::serialize(&self.old_hashchain)?;
+        self.inclusion_proof.verify_existence(self.old_root, self.key, old_serialized_hashchain)?;
+
+        let mut hashchain_after_update = self.old_hashchain.clone();
+        // Append the new entry and verify it's validity
+        hashchain_after_update.add_entry(self.new_entry.clone())?;
+
+        // Ensure the update proof corresponds to the new hashchain value
+        let new_serialized_hashchain = bincode::serialize(&hashchain_after_update)?;
+        self.update_proof.clone().verify_update(
+            self.old_root,
+            self.new_root,
+            vec![(self.key, Some(new_serialized_hashchain))],
+        )?;
+
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct MembershipProof {
+    pub root: Digest,
+    pub proof: SparseMerkleProof<Hasher>,
+    pub key: KeyHash,
+    pub value: Hashchain,
+}
+
+impl MembershipProof {
+    pub fn verify(&self) -> Result<()> {
+        let value = bincode::serialize(&self.value)?;
+        self.proof.verify_existence(self.root.into(), self.key, value)
+    }
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct NonMembershipProof {
+    pub root: Digest,
+    pub proof: SparseMerkleProof<Hasher>,
+    pub key: KeyHash,
+}
+
+impl NonMembershipProof {
+    pub fn verify(&self) -> Result<()> {
+        self.proof.verify_nonexistence(self.root.into(), self.key)
+    }
+}
